use uint8

+++ tests/exporter/test_mask_functions.py
@@ -10,7 +10,7 @@ from tests.exporter.util import draw_square_by_corner, draw_poly
 
 class TestMaskToObject2D(TestCase):
     def testExportOneSquare(self):
-        image = np.zeros([300, 200], dtype=np.int)
+        image = np.zeros([300, 200], dtype=np.uint8)
         image = draw_square_by_corner(image, 100, (150, 50), color=255)
 
         slices = mask_to_objects_2d(image)
@@ -20,7 +20,7 @@ class TestMaskToObject2D(TestCase):
         self.assertTrue(slices[0].polygon.equals(box(50, 150, 150, 250)), msg="Polygon is equal")
 
     def testOffset(self):
-        image = np.zeros([300, 200], dtype=np.int)
+        image = np.zeros([300, 200], dtype=np.uint8)
         image = draw_square_by_corner(image, 100, (150, 50), color=255)
 
         slices = mask_to_objects_2d(image, offset=(255, 320))
@@ -30,7 +30,7 @@ class TestMaskToObject2D(TestCase):
         self.assertTrue(slices[0].polygon.equals(box(305, 470, 405, 570)), msg="Polygon is equal")
 
     def testSeveralObjects(self):
-        image = np.zeros([300, 200], dtype=np.int)
+        image = np.zeros([300, 200], dtype=np.uint8)
         image = draw_square_by_corner(image, 50, (150, 50), color=255)
         image = draw_square_by_corner(image, 50, (205, 105), color=127)
 
@@ -45,7 +45,7 @@ class TestMaskToObject2D(TestCase):
         self.assertTrue(slices[1].polygon.equals(box(105, 205, 155, 255)), msg="Polygon is equal")
 
     def testMultipartPolygon(self):
-        image = np.zeros([300, 200], dtype=np.int)
+        image = np.zeros([300, 200], dtype=np.uint8)
         image = draw_square_by_corner(image, 50, (150, 50), color=255)
         image = draw_square_by_corner(image, 50, (201, 101), color=127)
 
@@ -58,7 +58,7 @@ class TestMaskToObject2D(TestCase):
         self.assertEqual(slices[1].label, 127)
 
     def testAdjacentWithoutSeperation(self):
-        image = np.zeros([300, 200], dtype=np.int)
+        image = np.zeros([300, 200], dtype=np.uint8)
         image = draw_square_by_corner(image, 50, (150, 50), color=255)
         image = draw_square_by_corner(image, 50, (150, 101), color=127)
 
@@ -71,7 +71,7 @@ class TestMaskToObject2D(TestCase):
         self.assertEqual(slices[1].label, 127)
 
     def testAdjacentWithSeparation(self):
-        image = np.zeros([300, 200], dtype=np.int)
+        image = np.zeros([300, 200], dtype=np.uint8)
         image = draw_square_by_corner(image, 50, (150, 50), color=255)
         image = draw_square_by_corner(image, 50, (150, 102), color=127)
 
@@ -86,7 +86,7 @@ class TestMaskToObject2D(TestCase):
         self.assertTrue(slices[1].polygon.equals(box(102, 150, 152, 200)), msg="Polygon is equal")
 
     def testSmallObject(self):
-        image = np.zeros([100, 100], dtype=np.int)
+        image = np.zeros([100, 100], dtype=np.uint8)
         image = draw_poly(image, Polygon([(15, 77), (15, 78), (16, 78), (15, 77)]), color=127)
         image = draw_poly(image, box(1, 1, 2, 2), color=255)
